Adding Authorization to the app

+++ backend/Routes/posts.js
@@ -32,9 +32,13 @@ router.post("",checkAuth, multer({storage:storage}).single("image"), (req, res, 
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath : url + "/images/" + req.file.filename
+    imagePath : url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   })
+
   console.log(post);
+  // console.log(req.userData);
+  // return res.status(200).json({});
   post.save().then(createdPost=>{
     res.status(201).json({
       message: "Post added succesfully",
@@ -71,9 +75,14 @@ router.get("", (req, res, next) => {
 });
 
 router.delete("/:id",checkAuth, (req, res, next)=>{
-  Post.deleteOne({_id: req.params.id}).then(result=>{
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result=>{
     console.log(result);
-    res.status(200).json({message: "Post Deleted!"});
+    if(result.n>0){
+      res.status(200).json({message: "Post Deleted!"});
+    }else{
+      res.status(401).json({message:"Not Authorized"});
+    }
+
   })
 });
 
@@ -98,11 +107,16 @@ router.put("/:id",checkAuth,multer({storage:storage}).single("image"),  (req, re
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath : imagePath
+    imagePath : imagePath,
+    creator: req.userData.userId
   });
   Post.updateOne({_id:req.params.id}, post).then(result=>{
     console.log(result);
-    res.status(200).json({message: "updated Sucessfullys!"});
+    if(result.nModified>0){
+      res.status(200).json({message: "updated Sucessfullys!"});
+    }else{
+      res.status(401).json({message:"Not Authorized"});
+    }
   });
 });
